+++ pages/user/[user].js
@@ -71,7 +71,11 @@ export default function User() {
               <div>{user.name}</div>
               <span>{user.login}</span>
             </h1>
-            {user.status && <div className={styles.status}>{parse(user.status.emojiHTML)} {user.status.message}</div>}
+            {user.status && (
+              <div className={styles.status}>
+                {parse(user.status.emojiHTML ?? '')} {user.status.message}
+              </div>
+            )}
           </div>
         </div>